Experimental: allow custom build script override
This serves as the temporary workaround for building MTR.
MTR is a multi-project gradle project: the correct task we need is
in a subproject, namely `:forge:remapJar`.
However, there is no easy way for us to locate a task in a subproject.
Further, MTR's buildscript also have copy tasks to rename the build
artifact, which adds up another complexity.
To overcome this without modify project's own script, I decide to
override our default build process.
In the future, we might want to investigate a better way to handle this.